First implementation of ziffect.
ziffect is a new library intended to provide some glue layers between zope
interfaces and Effect. It occurred to me while working with Effect that Effect
dispatchers are effectively classes, the types that they dispatch based upon
are effectively names of methods, and intents are effectively bundles of
arguments.

This is a first-pass proof of concept, that definitely took some shortcuts and
only implements the happy path for now. Future work will include tests for good
exceptions and actual attempts to use ziffect for something significant to see
if it is useful or not.